+++ src/pages/CommunityPage.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 
 import CommunityLists from '../components/Comunity/CommunityLists';
